Refactor is_function.hpp/is_member_function_pointer.hpp: With separate files for C++03 and C++11 versions. Add workarounds for compilers with insufficient C++11 support to handle the new code. Add support for C++17 noexcept function signatures and update tests accordingly.
